Replace okhttp with ktor
