+++ radar-jersey/src/main/kotlin/org/radarbase/jersey/service/managementportal/MPClientFactory.kt
@@ -1,24 +1,28 @@
 package org.radarbase.jersey.service.managementportal
 
-import com.fasterxml.jackson.databind.ObjectMapper
 import jakarta.ws.rs.core.Context
-import okhttp3.OkHttpClient
 import org.radarbase.jersey.auth.AuthConfig
+import org.radarbase.management.auth.ClientCredentialsConfig
+import org.radarbase.management.auth.clientCredentials
 import org.radarbase.management.client.MPClient
+import org.radarbase.management.client.mpClient
 import java.util.function.Supplier
 
 class MPClientFactory(
     @Context private val authConfig: AuthConfig,
-    @Context private val okHttpClient: OkHttpClient,
-    @Context private val objectMapper: ObjectMapper,
 ) : Supplier<MPClient> {
-    override fun get(): MPClient = MPClient(
-        serverConfig = MPClient.MPServerConfig(
-            url = requireNotNull(authConfig.managementPortal.httpUrl) { "ManagementPortal client needs a URL" }.toString(),
-            clientId = requireNotNull(authConfig.managementPortal.clientId) { "ManagementPortal client needs a client ID" },
-            clientSecret = requireNotNull(authConfig.managementPortal.clientSecret) { "ManagementPortal client needs a client secret" },
-        ),
-        objectMapper = objectMapper,
-        httpClient = okHttpClient,
-    )
+    override fun get(): MPClient = mpClient {
+        val mpUrl = requireNotNull(authConfig.managementPortal.url) { "ManagementPortal client needs a URL" }
+            .trimEnd('/') + '/'
+        auth {
+            clientCredentials(
+                ClientCredentialsConfig(
+                    tokenUrl = "$mpUrl/oauth/token",
+                    clientId = authConfig.managementPortal.clientId,
+                    clientSecret = authConfig.managementPortal.clientSecret,
+                ).copyWithEnv()
+            )
+        }
+        url = mpUrl
+    }
 }
